Use a global config for some parameters
This can later be loaded from JSON or something, but for now it just
lives with the rest of the code. I'm already going to mix data with code
in this repo and these values shouldn't change much, so I'm fine with
that.

This is my personal site after all. I'm not expecting to release this in
a way for others to use.